+++ src/main/java/com/onestack/project/service/AdminService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.onestack.project.domain.MemProPortCate;
 import com.onestack.project.domain.Member;
 import com.onestack.project.domain.Professional;
 import com.onestack.project.mapper.ManagerMapper;
@@ -21,12 +22,11 @@ public class AdminService {
 		return managerMapper.getAllMember();
 	}
 	
-	public List<Professional> getRequestPro(){
-		return managerMapper.getRequestPro();
+	public List<MemProPortCate> getMemProPortCate(){
+		return managerMapper.getMemProPortCate();
 	}
-	
-	public void updateProStatus(int proNo, Integer status, String screeningMsg) {
-	    managerMapper.updateProStatus(proNo, status, screeningMsg);
+	public void updateProStatus(int proNo, Integer professorStatus, String screeningMsg) {
+	    managerMapper.updateProStatus(proNo, professorStatus, screeningMsg);
 	}
 
 }
