+++ src/main/resources/static/js/admin/bootstrap-datepicker.js
