+++ src/main/resources/templates/adminDashboard/screeningManagement/reviewPendingInquiry.html
@@ -1,11 +1,8 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
-
     <title>전문가 심사 조회</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" >
-    <link href="/css/adminPage.css" rel="stylesheet" >
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+   <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js"></script>
 </head>
 <body>
     <!-- 전체를 감싸는 컨테이너 추가 -->
@@ -23,14 +20,14 @@
     							<span class="sort-down">▼</span>
 							</i>
                         </th>
-                        <th scope="col" class="sortable" data-sort="id">
+                        <th scope="col" class="sortable" data-sort="memberId">
                             아이디
 							<i class="sort-icon">
     							<span class="sort-up active">▲</span>
     							<span class="sort-down">▼</span>
 							</i>
                         </th>
-                        <th scope="col" class="sortable" data-sort="type">
+                        <th scope="col" class="sortable" data-sort="itemTitle">
                             전문 분야
 							<i class="sort-icon">
     							<span class="sort-up active">▲</span>
@@ -44,7 +41,7 @@
     							<span class="sort-down">▼</span>
 							</i>
                         </th>
-                        <th scope="col" class="sortable" data-sort="status">
+                        <th scope="col" class="sortable" data-sort="portfolioTitle">
                             포트폴리오
 							<i class="sort-icon">
     							<span class="sort-up active">▲</span>
@@ -52,7 +49,7 @@
 							</i>
                         </th>
                         <th scope="col" class="sortable" data-sort="date">
-                            요청일
+                            가입일
 							<i class="sort-icon">
     							<span class="sort-up active">▲</span>
     							<span class="sort-down">▼</span>
@@ -62,20 +59,20 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <tr th:each="pro : ${pro}">
+                    <tr th:each="pro : ${pro}" th:if="${pro.professional.professorStatus == 1}">
                         <td><input type="checkbox"/></td>
-                        <td th:text="${pro.name}">이름</td>
-                        <td th:text="${pro.memberId}">아이디</td>
-                        <td th:text="${pro.itemTitle}">전문 분야</td>
-                        <td th:text="${pro.email}">이메일</td>
-                        <td th:text="${pro.portfolioTitle}">포트폴리오</td>
-                        <td th:text="${#dates.format(pro.proDate, 'yyyy-MM-dd HH:mm:ss')}">요청일</td>
+                        <td th:text="${pro.member.name}">이름</td>
+                        <td th:text="${pro.member.memberId}">아이디</td>
+                        <td th:text="${pro.category.itemTitle}">전문 분야</td>
+                        <td th:text="${pro.member.email}">이메일</td>
+                        <td th:text="${pro.portfolio.portfolioTitle}">포트폴리오</td>
+                        <td th:text="${#dates.format(pro.member.memberRegDate, 'yyyy-MM-dd HH:mm:ss')}">회원 가입일</td>
                         <td>
 							<button class="btn btn-outline-dark btn-sm screening-btn" 
-							        th:data-pro-no="${pro.proNo}"
-							        th:data-name="${pro.name}"
-							        th:data-item-title="${pro.itemTitle}"
-							        th:data-portfolio-title="${pro.portfolioTitle}">
+							        th:data-pro-no="${pro.professional.proNo}"
+							        th:data-name="${pro.member.name}"
+							        th:data-item-title="${pro.category.itemTitle}"
+							        th:data-portfolio-title="${pro.portfolio.portfolioTitle}">
 							    승인 관리
 							</button>
                         </td>
@@ -107,14 +104,14 @@
     </div>
     
  <!-- 심사 관리 모달 -->
-<div class="modal fade" id="screeningModal" tabindex="-1" aria-labelledby="screeningModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-dialog-centered"> <!-- 중앙 정렬 추가 -->
-        <div class="modal-content border-0"> <!-- 테두리 제거 -->
-            <div class="modal-header border-bottom-0"> <!-- 헤더 테두리 제거 -->
-                <h5 class="modal-title fw-bold" id="screeningModalLabel">전문가 심사</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div class="modal-body px-4"> <!-- 패딩 증가 -->
+	<div class="modal fade" id="screeningModal" tabindex="-1" aria-labelledby="screeningModalLabel" aria-hidden="true">
+	    <div class="modal-dialog modal-dialog-centered">
+	        <div class="modal-content">
+	            <div class="modal-header">
+	                <h5 class="modal-title" id="screeningModalLabel">전문가 심사</h5>
+	                <button type="button" class="btn btn-outline-secondary close-custom-btn" data-modal-id="screeningModal">X</button>
+	            </div>
+	            <div class="modal-body px-4">
                 <form id="screeningForm">
                     <input type="hidden" id="proNo">
                     <div class="mb-4">
@@ -150,12 +147,11 @@
                 </form>
             </div>
             <div class="modal-footer border-top-0 px-4"> <!-- 푸터 테두리 제거 -->
-                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">취소</button>
+                <button type="button" class="btn btn-secondary close-custom-btn" data-modal-id="screeningModal">취소</button>
                 <button type="button" class="btn btn-primary" id="saveScreening">심사 완료</button>
             </div>
         </div>
     </div>
 </div>
-
 </body>
 </html>
