+++ src/main/resources/templates/layouts/admin_layout.html
@@ -32,10 +32,10 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/locales/bootstrap-datepicker.ko.min.js"></script>
     
     <!-- Custom Scripts -->
-    <script src="/js/adminDashboard.js"></script>
-    <script src="/js/datePicker.js"></script>
-    <script src="/js/adminArray.js"></script>
-    <script src="/js/adminFilter.js"></script>
-    <script src="/js/members.js"></script>
+    <script src="/js/admin/adminDashboard.js"></script>
+    <script src="/js/admin/datePicker.js"></script>
+    <script src="/js/admin/adminArray.js"></script>
+    <script src="/js/admin/adminFilter.js"></script>
+    <script src="/js/admin/members.js"></script>
 </body>
 </html>
